refactor: 장바구니 검증 수정
