+++ src/main/java/cart/domain/cartitem/CartItem.java
@@ -1,7 +1,7 @@
 package cart.domain.cartitem;
 
-import cart.domain.Member;
-import cart.domain.Product;
+import cart.domain.member.Member;
+import cart.domain.product.Product;
 import cart.exception.CartItemException;
 
 import java.util.Objects;
@@ -28,24 +28,26 @@ public class CartItem {
         this.member = member;
     }
 
-    public boolean isCorrectQuantity(final Long productId, final Integer quantity) {
-        if (this.quantity == quantity && product.isEqualId(productId)) {
-            return true;
-        }
-        throw new IllegalArgumentException("상품의 수량이 일치하지 않습니다.");
+    public boolean validateProduct(final Long productId, final Long memberId) {
+        return !(product.isEqualId(productId) && member.isEqualId(memberId));
     }
 
-    public boolean isNotOwnedByMember(final Member member) {
-        return !this.member.equals(member);
+    public boolean validateQuantity(final Integer quantity) {
+        return this.quantity != quantity;
     }
 
-    public void validate(final Long productId, final Long memberId, final Integer quantity) {
-        if (!(product.isEqualId(productId) && member.isEqualId(memberId))) {
-            throw new IllegalArgumentException("장바구니 정보가 일치하지 않습니다.");
+    public void checkOwner(Member member) {
+        if (!Objects.equals(this.member.getId(), member.getId())) {
+            throw new CartItemException.IllegalMember(this, member);
         }
-        if (this.quantity != quantity) {
-            throw new IllegalArgumentException("상품의 수량이 일치하지 않습니다.");
-        }
+    }
+
+    public void changeQuantity(int quantity) {
+        this.quantity = quantity;
+    }
+
+    public int calculatePriceOfQuantity() {
+        return product.calculatePriceBy(quantity);
     }
 
     public Long getId() {
@@ -70,20 +72,6 @@ public class CartItem {
 
     public int getQuantity() {
         return quantity;
-    }
-
-    public void checkOwner(Member member) {
-        if (!Objects.equals(this.member.getId(), member.getId())) {
-            throw new CartItemException.IllegalMember(this, member);
-        }
-    }
-
-    public void changeQuantity(int quantity) {
-        this.quantity = quantity;
-    }
-
-    public int calculatePriceOfQuantity() {
-        return product.calculatePriceBy(quantity);
     }
 
 }
